+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 6.3.0 (unreleased)
 
+- Add date_cancelled to User
 - Avoid logging in the history changes in the attachment's urls or its order (issue #tg-4247)
 
 ## 6.2.1 (2021-06-22)